+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Migrate `CustomDrugEntrySheet` to user view effects
 - Migrate `SettingsScreen` to use view effects
 - Bump Paging to v3.1.0
+- Migrate `RemoveOverdueAppointmentScreen` to user view effects
 
 ### Changes